requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pypdf==5.1.0
pillow==11.0.0
langgraph==0.2.34
langgraph-checkpoint-sqlite==2.0.1
requests==2.32.3
